Extract getLastArtboard function
